Added a function for making a new project
It doesn't get called anywhere in the frontend yet but I'll make it get
called once lncvrt is done with his changes to the frontend. Also for
now webserver and website both just have an index.php with one header.
But lncvrt will make it so it gets templates from a git repo.