Added a language packages menu to the project manager
What this will do is basically there will be a list of languages and u
will click install on any of them and then you will have auto complete
and syntax highlighting and stuff for that language. I don't want to
just have every lang that code fly supports automatically be part of
code fly.